fixup package repository url
The `repos` option contains URLs for package repositories. Entries can
optionally have names.

Package DESCRIPTION records include a `Repository` field that _may_ capture
where the package originated.

The package `Repository` value is something like CRAN or RSPM - these are
supplied by the download host. There is no guarantee that the `Repository`
corresponds to the (optional) names in `repos`.

This change uses `available.packages` to enumerate all packages known by the
`repos` option. We locate each package in the set of available packages and use
the contrib.url in that record to relate the package back to the configured
`repos` repository.

The package `Source` value is given the name of the `repos` repository. A
random name of the form "repo_XXXXXXXX" is assigned if the `repos` repository
is not named. The `Source` for CRAN packages is not altered.

The package `Source` name for non-CRAN packages is altered because we may have
multiple CRAN-like repositories in play, each with a different URL. We do not
want to have different URLs using the same `Source`.