adding purgeApp() function, updating POST() and terminateApp()
Adding a new function, purgeApp(), to access the
/applications/{id}/purge API end point on shinyapps.io.

This new function sends a request body with content length of zero. To
make this happen, the POST() function was updated to check if the
content and file variables are NULL. If so, a request with no body is
sent, otherwise the data is stored in the request's body. The old
behavior of the POST function was to raise an error if both of these
variables were NULL.

We care about not sending a request body because if there is no data to
send along with the request, we can either send no body or some form of
an empty JSON data type. API end points like
/applications/{id}/terminate don't expect a body, so we should not send
one. Previously, we were sending an empty list, but since the API didn't
look for a body, it did not care.  On the other hand, the
/applications/{id}/purge end point does accept data, and sending an
empty list will result in an error.

With the new ability to send POST requests without a body, the
terminateApp() function was also updated to not send a request body.

Adding test cases for changes to POST method