warn when unable to locate a package repository URL
Previously, we would err (and prohibit manifest creation and/or deployment)
when unable to locate a repository URL for a named package. Now, warn and
allow the operation to proceed.

Archived CRAN packages are not discoverable through available.packages.

Source-installed packages do not have repository URLs, but may have been
installed locally on the server (and configured as an external package
dependency).

Fixes #508