preserve file timestamps when copying files into bundle staging
This change makes preserves the timestamps from the source files when copying
into the temporary bundle staging directory. As a result, the tar.gz archive
will use those timestamps rather than the timestamps from the time the copy was
performed.

Having accurate timestamps within a bundle makes it possible for deployments to
make decisions based on the staleness of the deployed files. A pre-rendered R
Markdown document, for example, will have a newer timestamp when the included
HTML is stale. Prior to this change, the HTML and Rmd would almost always be
written in the same second and rmarkdown::run is unable to detect staleness.