Add a SHA-256 checksum to requests via the HTTP Digest header.
MD5 checksums are not all that secure, and cause problems on
FIPS-compliant systems in particular, so we'd ideally move to SHA-256.

In order to someday deprecate the use of MD5 while also continuing to
work with older versions of Connect that may require it, we have two
broad options:

1. Determine whether the upstream server supports another checksum
   mechanism, likely by checking its version; or

2. Keep the existing checksum support in place but send an additional
   non-MD5 checksum for use by servers that support it.

This commit implements the second approach: we send an additional header
with a SHA-256 checksum in addition to the existing MD5 one. Rather than
use another custom header, I've opted to use the Digest header specified
in RFC 3230, which is consistent with the semantics we're after anyway.

Unit tests are included but I have not done any end-to-end testing.

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>