Handle publishing from temp copy cloud sources (#638)
This fixes a bug that prevented creating published outputs from
temporary copies of projects in Posit Cloud. Ordinarily, when creating
an output from a project in cloud, the source and its output will be
linked in Posit Cloud, and the output will appear in the same space as
its source. In the case of temporary copy sources, we will allow
creating outputs, but do not link them to the temporary source project
and always publish the output to the author's personal workspace.

Co-authored-by: Mike Baynton <mike.baynton@posit.co>