Prepare for refactoring on `deploymentTarget()` (#662)
I've identified that `deploymentTarget()` is a large source of frustration for me, because when I misspecify the function arguments, it never gives me _quite_ enough information to immediately resolve the problem. This PR does not improve the errors, but prepares for a future PR that will do so. In this PR:

* I've moved `deploymentTarget()` to its own file, so it can have it's own test
* Extracted out `userNameFromDeployment()` and `createDeployment()` target into standalone functions
* Extract `findAppId()` out of `createDeploymentTarget()`.
* Added some mocking infrastructure
* Got to 100% test coverage of the main branches of `deploymentTarget()` (I have not yet tested the `findAppId()` path since that felt a bit peripheral to this PR)
* Added a bunch of TODOs that I would like to resolve in future PRs.

I've taken care to not change any behaviour in this PR, so any review can be more about the clarity of the tests and the mocking infrastructure. (Writing these tests have revealed a few annoyances in our mocking tools, which I'll tackle in the weeks to come.)