Implement `findAccount()`; simply `deploymentTarget()` (#666)
This PR introduces a new `findAccounts()` function that has _extensive_ messaging to help you when an account isn't found (this fixes a pet peeve of mine). This replaces `resolveAccount()` eliminating the `stopWith*()` family, adds some basic type checks, and uses `@inheritParams` to ensure all functions that use `account` and `server` have exactly the same documentation.

Using `findAccounts()` in `deploymentTargets()` started in on the road to a radical simplification. The logic is now clear, and based on the number of existing matching deployments (based on `appName`, `server`, and `account`). If there are 0 match, do an initial deploy; if there's 1 match, update the existing deploy; if there are more, error requesting that the user disambiguate. This also resolves some combinations that failed to use the previous `appTitle`, moves additional title logic out of `saveDeployment()`.