Mild refactoring of app metadata (#673)
Introduce a new function `appMetadata()` that collects the results of `inferQuartoInfo()`, `inferAppMode()`, `inferAppPrimaryDoc()`, `appHasParameters()`, and `detectPythonInDocuments()` into a single object. These functions (and their helpers) now all live a new, `appMetadata.R`, and I've increased the test coverage. I've done some small refactorings along the way to make the code a little clearer, but I've tried to stick to low-risk changes that are unlikely to impact the operation of the code.