Capture python metadata into a single object (#683)
Moves all python related code to a single file, and then extracts the python metadata (`python`, `condaMode`, and `forceGenerate`) into a single `pythonConfigurator()` object. This might be an overly clever approach, but the single object I chose to use is a closure, so we can lazily infer the python env as needed.

I also added a new helper, `pythonPathOrSkip()` (mimicking `quartoPathOrSkip()`) that uses reticulate to the path to python for tests.