Discover untracked deployment in deploymentForTarget() (#726)
This eliminates `applicationForTarget()` in favour of including all the logic in `deploymentForTarget()`, and means that the messaging can move back into `deployApp()`.

While I was in here, I also made the options for dealing with an update more clear, and used similar logic from `generateAppName(unique = TRUE)` to create a unique name if the user doesn't want to update. To make this interactive path testable, I sketched out some potential future cli functions, which for now live in `utils-cli.R`.

This eliminates one previous code path that I know @aronatkins cares about: what happens if an app has been deleted on the server. I need to do some research to double check what happened previously, but my gut feeling is that it wasn't handled gracefully, and we should tackle in a separate PR.