Better handling for deleted apps (#770)
If `deployApp()` determines that it's publishing to an app that doesn't exist, it offers the user the choice to create a new app. Fixes #226.

Includes somewhat of a revamp to `syncAppMetadata()` to delete deployments that don't correspond to known applications. Interactive testing of this function revealed some other problems that I fixed, along with generally tidying up the code.

New `addTestDeployment()` and `simulate_user_input()` to make testing easier.