Tweak handling when appId set
The IDE supplies both `appName` and `appId` when deploying so we can't assume supply `appId` means "create record from data on server". Now, we only take that path if `appId` is set but `appName` is not, and a check to make sure that if both are specified they consistent.

Fixes #775