prefer RSCONNECT_PACKRAT over rsconnect.packrat
Environment variables are visible in the RStudio IDE R session used for
push-button deployment. Not all R options are mirrored into that session.

These settings indicate that packrat should be used, even in the presence
of an renv.lock. When using packrat, the renv directory and renv.lock file
are removed from the bundle directory, as would happen prior to rsconnect 1.0.0.

Fixes #935
Fixes #936