appMode override
Allow override of the inferred `appMode` for situations where complicated
deployments can lead to incorrectly inferred content type. For example, the
presence of an `app.R` forces the content to be deployed as a Shiny
application with the `appMode = "shiny"` even if that file was an auxiliary
resource for some other type of content that happens to have lower precedence.

The appMode argument is accepted by:
- `deployApp`
- `writeManifest`
- `listDeploymentFiles`
- `appDependencies`

The `appDependencies` function supports obtaining a set of deployment files
through `appFileManifest`.

The `appMetadata` function no longer calls `listDeploymentFiles`. It assumes
that the caller has already fully enumerated the required files and provided
them in `appFiles`.

Fixes #948