create deployment-from-application helper
This rewrite showed that from-disk deployment records had different field
names than the records created on-the-fly. In particular, "appName" and
"appTitle" were used by the "deployment target", while "name" and "title" were
used for on-disk deployment records. This made it challenging to use the same
helpers.

All deployment objects, either read from disk or created in-memory now use
"name" and "title".

This is a step towards always using deployment records.