feat: make snowflakeConnectionName optional for some cases
This commit allows `addServer()`, `validateServerUrl()`, and
`connectSPCSUser()` to provide a default for the
`snowflakeConnectionName` parameter using the existing conventions for a
default Snowflake connection. This should simplify programmatic
deployment from RStudio, particularly in the Native App, where we can
assume a default connection is available.

The default connection is only used when it appears to match the account
used for the Connect server URL.

Unit tests are included.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>